+++ tests/OrderReturn/OrderReturnServiceTest/testGetOrderReturnsWithoutData.yml
@@ -16,14 +16,14 @@
             code: '200'
             message: OK
         headers:
-            Date: 'Tue, 12 Sep 2017 15:03:36 GMT'
+            Date: 'Wed, 20 Sep 2017 09:27:43 GMT'
             Server: 'Apache/2.4.10 (Debian)'
             Cache-Control: 'no-cache, private'
-            X-Debug-Token: c6cfa9
-            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/c6cfa9'
+            X-Debug-Token: e01f08
+            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/e01f08'
             Content-Length: '60'
             Content-Type: application/json
-        body: '{"id":7,"apiKey":"B8QJ3lwp7tDjumx3FC2wf6xhgnjYHVwAPBASJhGc"}'
+        body: '{"id":7,"apiKey":"2cbgA3pAqG4tNL1ht+CpFOdKP7mVC4F5LW+JpYUK"}'
 -
     request:
         method: GET
@@ -32,7 +32,7 @@
             Host: wizaplace.loc
             Accept-Encoding: null
             User-Agent: 'GuzzleHttp/6.2.1 curl/7.38.0 PHP/7.1.9'
-            Authorization: 'token B8QJ3lwp7tDjumx3FC2wf6xhgnjYHVwAPBASJhGc'
+            Authorization: 'token 2cbgA3pAqG4tNL1ht+CpFOdKP7mVC4F5LW+JpYUK'
             VCR-index: '1'
             Accept: null
     response:
@@ -41,11 +41,11 @@
             code: '200'
             message: OK
         headers:
-            Date: 'Tue, 12 Sep 2017 15:03:37 GMT'
+            Date: 'Wed, 20 Sep 2017 09:27:43 GMT'
             Server: 'Apache/2.4.10 (Debian)'
             Cache-Control: 'no-cache, private'
-            X-Debug-Token: f3cef0
-            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/f3cef0'
+            X-Debug-Token: 0e2e81
+            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/0e2e81'
             Content-Length: '2'
             Content-Type: application/json
         body: '[]'
