+++ tests/OrderReturn/OrderReturnServiceTest/testGetReturnReasons.yml
@@ -15,11 +15,11 @@
             code: '200'
             message: OK
         headers:
-            Date: 'Tue, 12 Sep 2017 15:04:07 GMT'
+            Date: 'Wed, 20 Sep 2017 09:40:36 GMT'
             Server: 'Apache/2.4.10 (Debian)'
             Cache-Control: 'no-cache, private'
-            X-Debug-Token: e49575
-            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/e49575'
+            X-Debug-Token: 26b7b2
+            X-Debug-Token-Link: 'http://wizaplace.loc/_profiler/26b7b2'
             Content-Length: '522'
             Content-Type: application/json
         body: '[{"id":3,"name":"Colis arriv\u00e9 en mauvais \u00e9tat, endommag\u00e9","position":10},{"id":4,"name":"Ne correspond pas au produit command\u00e9","position":20},{"id":5,"name":"Produit cass\u00e9, d\u00e9fectueux","position":30},{"id":6,"name":"Je n\u0027arrive pas \u00e0 le faire fonctionner ou \u00e0 le monter","position":40},{"id":7,"name":"Pas la bonne taille","position":50},{"id":8,"name":"Erreur de commande","position":60},{"id":9,"name":"Ne me convient plus, j\u0027ai chang\u00e9 d\u0027avis","position":70}]'
